fix(schema): restack provenance guidance on main
